+++ app/Http/Controllers/DepartmentController.php
@@ -4,15 +4,29 @@ namespace App\Http\Controllers;
 
 use App\Models\Department;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Yajra\DataTables\DataTables;
 
 class DepartmentController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('department.index')) {
+            abort(403, 'Unauthorized');
+        }
         $datas = Department::latest()->get();
         return view('admin.department.index', compact('datas'));
     }
@@ -22,6 +36,9 @@ class DepartmentController extends Controller
      */
     public function create()
     {
+        if (is_null($this->user) || !$this->user->can('department.create')) {
+            abort(403, 'Unauthorized');
+        }
         return view('admin.department.create');
     }
 
@@ -30,6 +47,9 @@ class DepartmentController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('department.store')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'department_code' => 'required|unique:departments',
             'department_desc' => 'required',
@@ -64,6 +84,9 @@ class DepartmentController extends Controller
      */
     public function update(Request $request, Department $department)
     {
+        if (is_null($this->user) || !$this->user->can('department.update')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'department_code' => "required|unique:departments,department_code,". "$department->id'",
             'department_desc' => 'required',
@@ -81,6 +104,9 @@ class DepartmentController extends Controller
      */
     public function destroy(Department $department)
     {
+        if (is_null($this->user) || !$this->user->can('department.destroy')) {
+            abort(403, 'Unauthorized');
+        }
         $department->delete();
         return back()->with('success', 'Delete successfully.');
     }
