+++ app/Http/Controllers/DeshboardMenuController.php
@@ -6,17 +6,30 @@ use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Session;
 use App\Models\DeshboardMenu;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
-use Auth;
 
 class DeshboardMenuController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('menu.index')) {
+            abort(403, 'Unauthorized');
+        }
         $datas = DeshboardMenu::orderBy('menu_short')->get();
         $perents = DeshboardMenu::orderBy('menu_short')->where('menu_perent',0)->select('menu_group', 'id')->get();
 
@@ -36,6 +49,9 @@ class DeshboardMenuController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('menu.store')) {
+            abort(403, 'Unauthorized');
+        }
 
         DeshboardMenu::create($request->except('_token') + [
             'userRole_id' => auth()->id(),
@@ -56,7 +72,7 @@ class DeshboardMenuController extends Controller
      */
     public function edit(DeshboardMenu $menu)
     {
-        
+
         $perents = DeshboardMenu::where('menu_perent',0)->select('menu_group', 'id')->get();
         return view('admin.dashboardMenu.edit', compact('menu', 'perents'));
     }
@@ -66,6 +82,9 @@ class DeshboardMenuController extends Controller
      */
     public function update(Request $request, DeshboardMenu $menu)
     {
+        if (is_null($this->user) || !$this->user->can('menu.update')) {
+            abort(403, 'Unauthorized');
+        }
         $menu_perent = DeshboardMenu::where('menu_group', $request->menu_group)->first();
         $menu->update($request->except('_token') + [
             'menu_perent' => $menu_perent->id
@@ -78,6 +97,9 @@ class DeshboardMenuController extends Controller
      */
     public function destroy(DeshboardMenu $menu)
     {
+        if (is_null($this->user) || !$this->user->can('menu.destroy')) {
+            abort(403, 'Unauthorized');
+        }
         $menu->delete();
         return back()->with('success', 'Successfully deleted.');
     }
