+++ app/Http/Controllers/GiroController.php
@@ -5,14 +5,28 @@ namespace App\Http\Controllers;
 use App\Models\Bank;
 use App\Models\Giro;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class GiroController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('giro.index')) {
+            abort(403, 'Unauthorized');
+        }
         $datas = Giro::latest()->get();
         $banks = Bank::latest()->get();
         return view('admin.giro.index', compact('datas', 'banks'));
@@ -31,6 +45,9 @@ class GiroController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('giro.store')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'giro_no' => 'required|integer',
             'bank_id' => 'required|integer',
@@ -62,6 +79,9 @@ class GiroController extends Controller
      */
     public function update(Request $request, Giro $giro)
     {
+        if (is_null($this->user) || !$this->user->can('giro.update')) {
+            abort(403, 'Unauthorized');
+        }
         $request->validate([
             'giro_no' => 'required|integer',
             'bank_id' => 'required|integer',
@@ -76,6 +96,9 @@ class GiroController extends Controller
      */
     public function destroy(Giro $giro)
     {
+        if (is_null($this->user) || !$this->user->can('giro.destroy')) {
+            abort(403, 'Unauthorized');
+        }
         $giro->delete();
         return back()->with('success', 'Deleted successfully.');
     }
