+++ app/Http/Controllers/MaritalStatusController.php
@@ -4,14 +4,29 @@ namespace App\Http\Controllers;
 
 use App\Models\maritalStatus;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class MaritalStatusController extends Controller
 {
+    public $user;
+
+
+    public function __construct()
+    {
+        $this->middleware(function ($request, $next) {
+            $this->user = Auth::guard('web')->user();
+            return $next($request);
+        });
+    }
+
     /**
      * Display a listing of the resource.
      */
     public function index()
     {
+        if (is_null($this->user) || !$this->user->can('marital-status.index')) {
+            abort(403, 'Unauthorized');
+        }
         $datas = maritalStatus::orderBy('marital_statuses_seqno')->get();
         return view('admin.maritalStatus.index', compact('datas'));
     }
@@ -29,6 +44,9 @@ class MaritalStatusController extends Controller
      */
     public function store(Request $request)
     {
+        if (is_null($this->user) || !$this->user->can('marital-status.store')) {
+            abort(403, 'Unauthorized');
+        }
 
         $request->validate([
             'marital_statuses_code' => 'required|unique:marital_statuses',
@@ -61,6 +79,9 @@ class MaritalStatusController extends Controller
      */
     public function update(Request $request, maritalStatus $marital_status)
     {
+        if (is_null($this->user) || !$this->user->can('marital-status.update')) {
+            abort(403, 'Unauthorized');
+        }
 
         $request->validate([
             'marital_statuses_code' => "required|unique:marital_statuses,marital_statuses_code,{$marital_status->id}",
@@ -78,6 +99,9 @@ class MaritalStatusController extends Controller
      */
     public function destroy(maritalStatus $marital_status)
     {
+        if (is_null($this->user) || !$this->user->can('marital-status.destroy')) {
+            abort(403, 'Unauthorized');
+        }
         $marital_status->delete();
         return back()->with('success', 'Deleted Successfully.');
     }
